+++ NEWS.md
@@ -1,5 +1,8 @@
 # lotri 0.2.2
 * Bug fix for conditional matrices
+* Now accessing `$lower` and `$upper` gives default values even if it
+  wasn't specified.
+* Can change the default conditional matrix to some other value like "id"
 
 # lotri 0.2.1
 * Added conditional matrix specification `|`; Returns a list of matrices